Split applications/new_schema.py to new folder structure
