+++ applications/new_schema/__init__.py
@@ -1,0 +1,17 @@
+from .mutations import Mutation
+from .queries import Query
+from .types import (
+    BerthApplicationFilter,
+    BerthApplicationNode,
+    BerthSwitchType,
+    HarborChoiceType,
+)
+
+__all__ = [
+    "BerthApplicationFilter",
+    "BerthApplicationNode",
+    "BerthSwitchType",
+    "HarborChoiceType",
+    "Mutation",
+    "Query",
+]
